Added stubs for the gRPC communication and moved all websocket comm to a new module.
