+++ kuksa_viss_client/KuksaGrpcComm.py
@@ -1,0 +1,17 @@
+#! /usr/bin/env python
+
+########################################################################
+# Copyright (c) 2022 Robert Bosch GmbH
+#
+# This program and the accompanying materials are made
+# available under the terms of the Eclipse Public License 2.0
+# which is available at https://www.eclipse.org/legal/epl-2.0/
+#
+# SPDX-License-Identifier: EPL-2.0
+########################################################################
+
+class KuksaGrpcComm:
+
+    # Constructor
+    def __init__(self, config):
+        pass
